requirements.txt
asgiref==3.11.0
bleach==6.3.0
certifi==2025.11.12
charset-normalizer==3.4.4
cloudinary==1.44.1
dj-database-url==3.0.1
Django==4.2.26
django-allauth==65.13.1
django-cloudinary-storage==0.3.0
django-summernote==0.8.20.0
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.11
packaging==25.0
psycopg2==2.9.11
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
whitenoise==6.11.0
